Removed module_loadex files, and moved their code inside commands'
